Updated Message Payload

+++ backend/graphql/src/resolvers/cart/handlers.rs
@@ -14,9 +14,9 @@ pub(crate) async fn add_cart_item(cart_item: NewCart) -> Result<Vec<Cart>, GqlEr
     .await?;
     let response = client
         .create_cart_item(CreateCartItemRequest {
-            product_id: cart_item.product_id.parse::<i32>().expect("err"),
-            quantity: cart_item.quantity.parse::<i32>().expect("err"),
-            user_id: cart_item.user_id.parse::<i32>().expect("err"),
+            product_id: cart_item.product_id.parse::<i64>().expect("err"),
+            quantity: cart_item.quantity.parse::<i64>().expect("err"),
+            user_id: cart_item.user_id.parse::<i64>().expect("err"),
         })
         .await?;
